Client Management Module + Bug Fixing
This were being implemented in admin panel, change from sweet alert to normal modal in order to get seemless design accross the system and reduce the third party things in the site, adding client first time login, altering the curent condition during the login for client to get into the first time login. Next is for booking and change the alert to tasker panel as well.